services/Sphana.Trainer/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "sphana_trainer"
version = "0.1.0"
description = "Training, evaluation, and ONNX export CLI for the Sphana NRDB."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Sphana Team"}]
keywords = ["rag", "knowledge-graph", "onnx", "training", "gnn"]
dependencies = [
    "typer[all]>=0.12.3",
    "rich>=13.7",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "hydra-core>=1.3",
    "loguru>=0.7",
    "onnx>=1.16",
    "onnxruntime-gpu>=1.20",
    "optimum[onnxruntime-gpu]>=1.20",
    "onnxscript>=0.1.0",
    "torch>=2.3",
    "transformers>=4.44",
    "sentence-transformers>=3.0",
    "fastembed>=0.2",
    "mlflow>=2.14",
    "numpy>=1.26",
    "scikit-learn>=1.5",
    "tqdm>=4.66",
    "sentencepiece>=0.2",
    "jsonschema>=4.19",
    "requests>=2.32"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-mock>=3.14",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.10"
]

[project.scripts]
sphana_trainer = "sphana_trainer.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/sphana_trainer"]
include = [
    "src/sphana_trainer/schemas/**/*.json",
]

[tool.hatch.build.targets.sdist]
include = [
    "src/sphana_trainer/schemas/**/*.json",
]

[tool.pytest.ini_options]
markers = [
    "ingestion: ingestion pipeline integration tests",
]


